Lineage: remove circular deps between models.ts and utils.ts
